+++ sunshine/View/Representable/VisualEffectViewRepresentable.swift
@@ -1,0 +1,19 @@
+//
+// Created by Michael Rippe on 1/28/21.
+//
+
+import SwiftUI
+
+struct VisualEffectView: UIViewRepresentable {
+
+    var effect: UIVisualEffect?
+
+    func makeUIView(context: UIViewRepresentableContext<Self>) -> UIVisualEffectView {
+        UIVisualEffectView()
+    }
+
+    func updateUIView(_ uiView: UIVisualEffectView, context: UIViewRepresentableContext<Self>) {
+        uiView.effect = effect
+    }
+
+}
